Fixing the problems with groupwise.wsdl, step 8: support for XML attributes.
This needed a major redesign of KDSoapValue/KDSoapValueList.
KDSoapValue can now hold both a value, and children (elements and attributes, in a KDSoapValueList)
This reduces the amount of qvariant-to-KDSoapValueList conversions too.

Added unittests for attributes.
Splitted out unittests based on sugarcrm.wsdl so that "builtinhttp" can remain dynamic-only.
Added code for sending, parsing, and serializing+deserializing (in the wsdl compiler) xml attributes.

This commit also makes the sending of xsi:type more logical: 1) only sent in EncodedUse mode, even for basic types,
2) passed as parameter rather than guessed, which sends better type names for xmlschema-defined specializations of basic types.


git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/products/kdsoap/trunk@36592 c66179b2-ff0d-0410-8aef-1b1a91e65cfa