Fix servertest with Qt-4.8.
The assumption that client sockets remain connected until we shutdown the server just doesn't seem
to be true anymore. So instead, count the number of connections in total, not the number of sockets
"connected right now".

Also the use of threads inside QNAM since 4.8 requires more file descriptors than before,
so take that into account in the rlimits.


git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/products/kdsoap/trunk@44390 c66179b2-ff0d-0410-8aef-1b1a91e65cfa