Fix conflict between complex type "Foo" and anonymous complex type "foo"
(named after the element <foo>), both leading to a C++ class "Foo".
This is done in multiple steps:
1) mark anonymous complex types as such
2) detect conflicts in TypeCollector, and mark the lowercase type as
conflicting
3) append a suffix (e.g. "Element", for anonymous types) to the C++ class
name, in TypeMap::addSchemaTypes.
This leaves the QName unchanged (so that type lookups work), and only
renames anonymous types in case of conflicts, to avoid breaking source
compatibility (and expectatiions) for all other anonymous types.

Testcase: TNS__DescribeLayout in salesforce-partner.wsdl.