Generate a ConfigVersion.cmake file for the cmake package.
As the macro used to write the file is new in CMake 2.8.8, require
that CMake version.

Also remove the old VERSION variables. The ConfigVersion.cmake file
defines the variables KDSoap_VERSION_MAJOR, KDSoap_VERSION_MINOR,
KDSoap_VERSION_PATCH and KDSoap_VERSION, so no need to add equivalents
in the Config.cmake file. Especially as the equivalents are oddly and
inconsistently named (KDSOAP_VERSION_VERSION_MINOR), I think this is just
a bug fix.