Fix the issue38 : replacing "." with "_"
Add a forbidden char in the className function of Style.cp file
Filter the classname given as parameter to be able to treat namespace and
type name, so the replacement of character is working.
Add an overload function of className to work with list of string
Change the results of three functions generating code in typemap
the return if filtered (if not buit in type) through class name
in order to have the class name,the forward declarations and references
to the class without "." and "_" instead.