Implement support for inheritance between complex types.
The value-based API (using const BaseType&) is unchanged, the magic happens
internally, with a QSharedPointer<BaseType> as storage.
Client-side it's all transparent.
Server-side, there's no way to know which derived type to instanciate, so
a getter for a raw KDSoapValue has been added, for dynamic introspection.

The case of list of elements (maxOccurs > 1), stored as QList< QSharedPointer<BaseType> >,
made this even more interesting; this case compiles now, but misses the KDSoapValue accessor.

issue26, SOAP-87, SOAP-62