Fix the issue SOAP-39 : import definition support.
Create the test with 3 wsdl, the main that does 2 imports, one of the
wsdl definition to include a binding and the other one to include the
xsd schema. The imported definition is solved by recursivity of
the method Definitions::loadXML. Logic of automatic service
creation has been moved into a method of Definition
be called by Compiler::parse.