Fix infinite recursion when type A contains a member A with minoccurs=0
It needs to be stored as a QSharedPointer rather than as a value,
otherwise the constructor recurses infinitely.

So the code now has two reasons to use QSharedPointer: because the
type of polymorphic or because a class contains an optional instance of itself.

An indirect composition (A contains B contains A) wouldn't be detected, though...

Testcase: encapsecurity/authstateless.wsdl (class __StatusCodeType)