Try harder to find a prefix for some namespaces.
This fixes some "__Class" without prefix, and it fixes the empty-prefix-clash
in encapsecurity. This isn't perfect yet though, the prefix "tns" is often
used for varying namespaces over time, so it all collides there.
We might really have to end up with a ns->prefix config file if we can't
figure it all automatically.