add common header and multiple class support
Some headers contain multiple classes per header, so if the entry is a
comma-separated string, then assume that the first string is the base name
and the following are additional classes that include the same original
header file.
At the end, also add a combined header if COMMON_HEADER is set.