don't abort the http call if authentication is not possible
If no credentials is available or if the credentials are incorrect,
then the real error message is hidden behind the cancellation of the
http call. By not aborting the call, the error state will stay
QNetworkReply::AuthenticationRequiredError. This will result is KDSoap
generating a SoapFault containing the correct error message.
This way the application can detect the error and ask the user for
correct credentials.