Add support for soap 1.2 fault in faultAsString
Soap 1.2 has a different fault format then soap 1.1. This can be
detected by the namespace that the fault is in. In soap 1.2 faults are
identified by a machine-readable error code and this can be further
categorized with subcodes. It contains human-readable reason is send as
well. This change allows faultAsString to show soap 1.2 faults in a
similar way as soap 1.1 faults.