Improve detection of soap fault in http response
The soap 1.2 spec states that soap faults of type env:Sender should have
http status code 400 Bad request. The previous code assumed that all
soap error have status code 500 (which is true for soap 1.1). This
change tries to find a soap fault in the response data (regardless of
the http status code) and will return that if available. Else it will
fallback to the generic fault based on the response error.

https://github.com/KDAB/KDSoap/pull/129