kdwsdl2cpp: Added import-path argument to search for files locally
Normally files needed by kdwsdl2cpp are downloaded. This patch adds an
option to search for the files locally. This increases the compilation
speed and improved reproducibility.
When the import-path is supplied, the needed files (for example schemas)
are search at: <importpath>/<URL-host>/<URL-path>
If the file is not found at the import path, then it will be downloaded
as it normally would.
For example the schema http://www.w3.org/2001/xml.xsd with the option
"-import-path /tmp" will be search at /tmp/www.w3.org/2001/xml.xsd.
The benefit is that the needed schemas can be pre-downloaded to increase
compilation speed or can be supplied with the wsdl to increase
reproducibility.