fix spec compliance
as per https://www.w3.org/Submission/ws-addressing/ the only well-known
uri in 2004/08 was anonymous and it has a role prefix. from the
implicit reading I can find on earlier versions the same applies to them.

so, when constructing the well-known URIs inject a role prefix when
dealing with specs up to 2004/08 and raise a warning when attempting to
use anything but anonymous

same for /id/unspecified