kdwsdl2cpp: Avoid potential type collisions in nested complexTypes
Manual merge of pull request from Holger Kaelberer in KDSoap since
the code moved to this repository.

So far, for complexTypes nested in elements with the same name a
single C++ class has been created. This is wrong in cases where
multiple elements with the same name contain complexTypes of different
structure because different classes should be created

Beside, new types for structurally different complexTypes only.