Fix generated code for an enumeration type with a length restriction (#234)
The existing unittest list_restriction didn't catch that things actually compiled.
So I removed it and extended the other unittests instead, to try and
test all cases (restriction on enum, restriction on string, restriction
on number of items in a list...)

The actual bugfix involves the addition of a stringValue() method for
enumeration types -- which also simplifies the serialize() code.